requirements.txt
fastapi
uvicorn
pydantic
requests
dotenv
python-multipart
supabase
piper-tts
google-cloud-speech
jinja2
pydub
tqdm